+++ src/basicfilters.rs
@@ -1,5 +1,6 @@
 use std::sync::{Arc, RwLock};
 
+use crate::audiodevice::AudioChunk;
 use crate::biquad::{Biquad, BiquadCoefficients};
 use crate::config;
 use crate::fifoqueue::FifoQueue;
@@ -35,9 +36,11 @@ pub struct Volume {
     samplerate: usize,
     chunksize: usize,
     processing_status: Arc<RwLock<ProcessingParameters>>,
+    fader: usize,
 }
 
 impl Volume {
+    #[allow(clippy::too_many_arguments)]
     pub fn new(
         name: String,
         ramp_time_ms: f32,
@@ -46,6 +49,7 @@ impl Volume {
         chunksize: usize,
         samplerate: usize,
         processing_status: Arc<RwLock<ProcessingParameters>>,
+        fader: usize,
     ) -> Self {
         let ramptime_in_chunks =
             (ramp_time_ms / (1000.0 * chunksize as f32 / samplerate as f32)).round() as usize;
@@ -68,6 +72,7 @@ impl Volume {
             samplerate,
             chunksize,
             processing_status,
+            fader,
         }
     }
 
@@ -78,8 +83,9 @@ impl Volume {
         samplerate: usize,
         processing_status: Arc<RwLock<ProcessingParameters>>,
     ) -> Self {
-        let current_volume = processing_status.read().unwrap().volume;
-        let mute = processing_status.read().unwrap().mute;
+        let fader = conf.fader as usize;
+        let current_volume = processing_status.read().unwrap().target_volume[fader];
+        let mute = processing_status.read().unwrap().mute[fader];
         Volume::new(
             name,
             conf.get_ramp_time(),
@@ -88,6 +94,7 @@ impl Volume {
             chunksize,
             samplerate,
             processing_status,
+            fader,
         )
     }
 
@@ -112,16 +119,10 @@ impl Volume {
             })
             .collect()
     }
-}
 
-impl Filter for Volume {
-    fn name(&self) -> String {
-        self.name.clone()
-    }
-
-    fn process_waveform(&mut self, waveform: &mut [PrcFmt]) -> Res<()> {
-        let shared_vol = self.processing_status.read().unwrap().volume;
-        let shared_mute = self.processing_status.read().unwrap().mute;
+    fn prepare_processing(&mut self) {
+        let shared_vol = self.processing_status.read().unwrap().target_volume[self.fader];
+        let shared_mute = self.processing_status.read().unwrap().mute[self.fader];
 
         // Volume setting changed
         if (shared_vol - self.target_volume).abs() > 0.01 || self.mute != shared_mute {
@@ -157,6 +158,49 @@ impl Filter for Volume {
             };
             self.mute = shared_mute;
         }
+    }
+
+    pub fn process_chunk(&mut self, chunk: &mut AudioChunk) {
+        self.prepare_processing();
+
+        // Not in a ramp
+        if self.ramp_step == 0 {
+            for waveform in chunk.waveforms.iter_mut() {
+                for item in waveform.iter_mut() {
+                    *item *= self.target_linear_gain;
+                }
+            }
+        }
+        // Ramping
+        else if self.ramp_step <= self.ramptime_in_chunks {
+            trace!("ramp step {}", self.ramp_step);
+            let ramp = self.make_ramp();
+            self.ramp_step += 1;
+            if self.ramp_step > self.ramptime_in_chunks {
+                // Last step of ramp
+                self.ramp_step = 0;
+            }
+            for waveform in chunk.waveforms.iter_mut() {
+                for (item, stepgain) in waveform.iter_mut().zip(ramp.iter()) {
+                    *item *= *stepgain;
+                }
+            }
+            self.current_volume = 20.0 * ramp.last().unwrap().log10();
+        }
+
+        // Update shared current volume
+        self.processing_status.write().unwrap().current_volume[self.fader] =
+            self.current_volume as f32;
+    }
+}
+
+impl Filter for Volume {
+    fn name(&self) -> String {
+        self.name.clone()
+    }
+
+    fn process_waveform(&mut self, waveform: &mut [PrcFmt]) -> Res<()> {
+        self.prepare_processing();
 
         // Not in a ramp
         if self.ramp_step == 0 {
@@ -178,6 +222,10 @@ impl Filter for Volume {
             }
             self.current_volume = 20.0 * ramp.last().unwrap().log10();
         }
+
+        // Update shared current volume
+        self.processing_status.write().unwrap().current_volume[self.fader] =
+            self.current_volume as f32;
         Ok(())
     }
 
@@ -189,6 +237,7 @@ impl Filter for Volume {
             self.ramptime_in_chunks = (conf.get_ramp_time()
                 / (1000.0 * self.chunksize as f32 / self.samplerate as f32))
                 .round() as usize;
+            self.fader = conf.fader as usize;
         } else {
             // This should never happen unless there is a bug somewhere else
             panic!("Invalid config change!");